fix warning

+++ src/client/utils/log.c
@@ -51,10 +51,7 @@ int xdag_log(int level, int err, char* file, int line, const char *format, ...)
 
 	int pos = 0;
 	char buffer[4096] = {0};
-
-//	pos = sprintf(buffer, "%s.%03d [%012llx:%.4s]  ", tbuf, (int)(tv.tv_usec / 1000), (long long)pthread_self_ptr(), lvl + 4 * level);
-	//// pos += sprintf(buffer, "[%012llx][%.4s][%s] %s:%d  ",  (long long)pthread_self(), lvl + 4 * level, tbuf, xdag_filename(file), line);
-	pos += sprintf(buffer, "[%012llx][%.4s][%s] %s:%d  ", pthread_self(), lvl + 4 * level, tbuf, xdag_filename(file), line);
+	pos += sprintf(buffer, "[%012llx][%.4s][%s] %s:%d  ", (long long)pthread_self(), lvl + 4 * level, tbuf, xdag_filename(file), line);
 	
 	va_start(arg, format);
 	pos = vsprintf(buffer + pos, format, arg);
